fix(cli): let trust + onboard see the non-workspace publishables
`gjsify onboard` reported "127 package(s) | 0 to publish+trust" while
`@gjsify/napi` was neither in the list nor published at all. It could not
be: `discoverWorkspaces` only sees ROOT WORKSPACES, and the packages under
`packages/{node-gi,napi}/*` deliberately are not — each is an independent
native engine with its own CI workflow, kept out of the workspace graph so
a full `gjsify foreach` never tries to build it.

`gjsify trust` had a hand-rolled carve-out for `packages/node-gi/*` only.
`gjsify onboard` had none at all, so it was blind to FOUR published
packages, not one:

  onboard before: 127 -> after: 131
  newly visible: @gjsify/gtk-runtime-darwin-arm64,
                 @gjsify/gtk-runtime-win32-x64,
                 @gjsify/node-gi,
                 @gjsify/napi

Both commands now share `utils/publishable-packages.ts`, whose directory
list is the single place to extend when a new non-workspace publishable
group appears. Getting that wrong is not cosmetic: the sweep reports
"all done", nobody configures a Trusted Publisher, and the package's first
release fails the OIDC exchange with `404 — package not found` — which
then breaks the serialized publish loop for everything after it.

The two `@gjsify/example-node-gi-*` packages under that directory are
`private: true` and stay filtered out, so nothing newly publishable
appears that should not be.
